Making src/Mint-Y/xfwm4-src/ more standard, like the others.
	modified:   src/Mint-Y/xfwm4-src/assets-dark.svg
	modified:   src/Mint-Y/xfwm4-src/assets.svg
	modified:   src/Mint-Y/xfwm4-src/render-assets.sh
	modified:   src/Mint-Y/xfwm4-src/xfwm4-dark/close-active.xpm
	modified:   src/Mint-Y/xfwm4-src/xfwm4-dark/close-prelight.xpm
	modified:   src/Mint-Y/xfwm4-src/xfwm4-dark/close-pressed.xpm
	modified:   src/Mint-Y/xfwm4-src/xfwm4/close-active.xpm
	modified:   src/Mint-Y/xfwm4-src/xfwm4/close-prelight.xpm
	modified:   src/Mint-Y/xfwm4-src/xfwm4/close-pressed.xpm
	modified:   update-variations.py

+++ src/Mint-Y/xfwm4-src/render-assets.sh
@@ -11,24 +11,27 @@ INDEX="assets.txt"
 for i in `cat $INDEX`
 do
 
-rm -f $ASSETS_DIR/$i.xpm
-echo
-echo Rendering $ASSETS_DIR/$i.png and converting to xpm
-$INKSCAPE --export-id=$i \
+if [ -f $ASSETS_DIR/$i.xpm ]; then
+    echo $ASSETS_DIR/$i.xpm exists.
+else
+    echo
+    echo Rendering $ASSETS_DIR/$i.png and converting to xpm
+    $INKSCAPE --export-id=$i \
           --export-id-only \
           --export-filename=$ASSETS_DIR/$i.png $SRC_FILE >/dev/null \
-&& convert $ASSETS_DIR/$i.png $ASSETS_DIR/$i.xpm
-rm -f $ASSETS_DIR/$i.png
-
-rm -f $DARK_ASSETS_DIR/$i.xpm
-echo
-echo Rendering $DARK_ASSETS_DIR/$i.png and converting to xpm
-$INKSCAPE --export-id=$i \
+    && convert $ASSETS_DIR/$i.png $ASSETS_DIR/$i.xpm
+    rm -f $ASSETS_DIR/$i.png
+fi
+if [ -f $DARK_ASSETS_DIR/$i.xpm ]; then
+    echo $DARK_ASSETS_DIR/$i.xpm exists.
+else
+    echo
+    echo Rendering $DARK_ASSETS_DIR/$i.png and converting to xpm
+    $INKSCAPE --export-id=$i \
           --export-id-only \
           --export-filename=$DARK_ASSETS_DIR/$i.png $DARK_SRC_FILE >/dev/null \
-&& convert $DARK_ASSETS_DIR/$i.png $DARK_ASSETS_DIR/$i.xpm
-rm -f $DARK_ASSETS_DIR/$i.png
-
-echo
+    && convert $DARK_ASSETS_DIR/$i.png $DARK_ASSETS_DIR/$i.xpm
+    rm -f $DARK_ASSETS_DIR/$i.png
+fi
 done
 exit 0
